==> 4.hrdsum_calculation.Rmd <==
---
title: "HRDsum scores"
output: github_document
---

```{r setup0, echo=FALSE, eval=TRUE, message=FALSE, warning=FALSE}
library(tidyverse)
library(yaml)
library(CHORD)
library(BSgenome.Hsapiens.UCSC.hg38)
library(RColorBrewer)
library(here)
source("scripts/plot_utils.R")
source("scripts/hrdsum.R")
load("data/hrdsum_chrominfo.rda")
config <- read_yaml("config.yaml")
extra_tumours <- config$extra_tumours
knitr::opts_chunk$set(root.dir = here(), comment = NA, echo = FALSE, message = FALSE, warnings = FALSE, include = TRUE, fig.align = 'center', fig.retina = 2, fig.path = "figures/", cache.path ="cache/")
```


The HRDsum scores were calculated using the `scar_score()` function of [modified scripts](scripts/hrdsum.R) based on scarHRD package. 


```{r, cache=TRUE, eval=FALSE}
run_hrdsum <- function(cnv_file) {
  sample <- basename(cnv_file) |> str_remove(".cnv.txt")
  read.table(cnv_file,header=T, check.names = F, stringsAsFactors = F, sep="\t") |> 
    select(Chromosome, chromStart, chromEnd, total.copy.number.inTumour, minor.copy.number.inTumour) |> 
    mutate(cn_B=total.copy.number.inTumour-minor.copy.number.inTumour, Chromosome=paste0("chr",Chromosome)) |> 
    add_column(ploidy=0, sample=sample) |> 
    select(sample,everything()) |> 
    scar_score() |> 
    as.data.frame()
}

df_hrdsum <- list.files("data/GRCh38", pattern = "*.cnv.txt", recursive = TRUE, full.names = TRUE) |> set_names(\(x) basename(x) |> str_remove(".cnv.txt")) |> 
  map(run_hrdsum) |>  list_rbind(names_to = "sample")

df_hrdsum |> write_tsv("results/hrdsum_355.tsv")
```


```{r}
df_hrdsum <- read_tsv("results/chord_hrdprediction_355.tsv") |>
  left_join(read_tsv("results/hrdsum_355.tsv"), by="sample") |> 
  mutate(sample_id=case_when(startsWith(sample,"MAGIC")~str_split_i(sample,"\\.",i=1),
                             startsWith(sample,"GQ")~str_split_i(sample,"\\.",i=1),
                             TRUE~sample),
         tumour_id=case_when(startsWith(sample,"MAGIC")~str_split_i(sample,"\\.",i=2),
                             startsWith(sample,"GQ")~str_split_i(sample,"\\.",i=2),
                             TRUE~sample)) |> 
  filter(!tumour_id %in% extra_tumours) |> 
  mutate(cohort=case_when(startsWith(sample,"MAGIC")~"MAGIC",
                          startsWith(sample,"TCGA")~"TCGA-BRCA",
                          startsWith(sample,"FBC")~"Familial Breast",
                          startsWith(sample,"GQ")~"Q-IMPROvE")) |> 
  arrange(desc(`HRD-sum`)) |> 
  mutate(sample_id=factor(sample_id,levels=sample_id), 
         hr_status=ifelse(hr_status=="HR_deficient","HRD", "HRP"), 
         cohort=factor(cohort, levels=cohort_levels))
```

There is no threshold for HRDsum score for distinguishing HRD samples from non-HRD samples.

```{r hrdsum_fig1, fig.width=4, fig.height=4}
p_hrdsum_threshold <- df_hrdsum |> 
  ggplot(aes(x=sample_id, y=`HRD-sum`)) +
  geom_point(alpha=.6, color="#3B4992") + 
  geom_hline(yintercept = 42, color="grey", lty=2, linewidth=1) +  annotate("text", x=260, y=50, label="cutoff=42", size=3.6, color="black") +
  theme_Publication() + labs(x="Tumor samples",y=stringr::str_wrap("HRDsum score", 10)) + 
  theme(legend.position = "bottom", axis.text.x = element_blank(), axis.ticks.x = element_blank())

p_hrdsum_threshold
```

**Figure 1: The HRDsum scores for tumour samples**

We did see higher HRDsum scores for HRD samples (CHORD) than non-HRD samples, but there are many overlapping points especially other cohorts than those Familial breast cancers.

```{r hrdsum_fig2, fig.width=7, fig.height=4.5, warning=FALSE}
df_hrdsum |> 
  ggplot(aes(x=hr_status, y=`HRD-sum`, color=hr_status)) + 
  scale_color_manual(values=c("#0073C2","#EFC000")) +
  geom_boxplot() + facet_grid(~cohort) + 
  geom_jitter(position=position_jitter(0.2)) +
  geom_hline(yintercept = 42, linetype=2 ,size=1, color="darkgrey") +
  theme_Publication() + 
  labs(x="", y="HRDsum score", color="HR status (CHORD)")
```

**Figure 2: The HRDsum scores for tumour samples in four cohorts with HRD and non-HRD predicted by CHORD. The grey dash line represent the cut-off value**

```{r hrdsum_fig3, fig.height=4, fig.width=4}
df_hrdsum |> 
  ggplot(aes(x=hr_status, y=`HRD-sum`, color=hr_status)) + 
  scale_color_manual(values=c("#0073C2","#EFC000")) +
  geom_boxplot() +
  geom_jitter() +
  geom_hline(yintercept = 42, linetype=2 ,size=1, color="darkgrey") +
  theme_Publication() + 
  labs(x="", y="HRDsum score", color="HR status (CHORD)")
```

**Figure 3: The HRDsum scores for tumour samples  with HRD and non-HRD predicted by CHORD. The grey dash line represent the cut-off value**

Although there is no threshold for HRDsum scores. HRDsum with a cutoff point of 42 is an FDA‐ and European Medicines Agency (EMA)‐approved biomarker to select ovarian cancer patients for PARP inhibition. 

Here we can see using a cut-off of 42 will predict a lot of non-HRD (identified by CHORD) as HRD.

```{r hrdsum_fig4,fig.width=7, fig.height=4.5, warning=FALSE}
df_hrdsum |> mutate(hr_status=ifelse(`HRD-sum`>=42,"HRD","HRP")) |> 
  ggplot(aes(x=hr_status,fill=hr_status)) + geom_bar() + 
  geom_text(stat='count', aes(label=..count..), vjust=-.1) +
  facet_grid(~cohort) + ylim(0,130) +
  theme_Publication() + labs(x="", y= "Number of samples", fill="HR status") + scale_fill_manual(values=c("#0073C2","#EFC000")) +
  theme(axis.text.x = element_blank(), axis.ticks.x = element_blank())
```

**Figure 4: The number of individuals identified as HRD or non-HRD in four cohorts by HRD-sum score with a cut-off point of 42**
